Add samples to bookshop (hash)

+++ src/fr/uvsq/zumu_panchi/model/Cart.java
@@ -27,9 +27,9 @@ public class Cart {
 
 	// ----------- << method.annotations@AAAAAAFxcAr4XHz/LIY= >>
 	// ----------- >>
-	public int getPrice() {
+	public float getPrice() {
 		// ----------- << method.body@AAAAAAFxcAr4XHz/LIY= >>
-		int totalPrice;
+		float totalPrice;
 
 		totalPrice = 0;
